restructure api-gateway
